+++ webapi/apiv1/resources/dataset.py
@@ -36,12 +36,8 @@ class DatasetResource:
         logger.debug('Authorize user {id}'.format(id=user_id))
 
         try:
-            if user_id:
-                query = Q(id=id) & (Q(base__owner=user_id) | Q(is_public=True))
-                dataset_meta = metadata.DatasetMetadata.from_id(query)
-            else:
-                kwargs = {'id': id, 'is_public': True}
-                dataset_meta = metadata.DatasetMetadata.from_id(**kwargs)
+            context = {'user_id': user_id}
+            dataset_meta = manager.get_dataset(id, context)
         except metadata.DoesNotExist:
             logger.debug('Dataset {id} does not exist'.format(id=id))
 
@@ -52,7 +48,7 @@ class DatasetResource:
 
         resp.status = falcon.HTTP_200
         dataset_meta_dict = dataset_meta.to_dict()
-        result_keys = ['id', 'status', 'is_public', 'title', 'description', 'category', 'labels']
+        result_keys = ['id', 'status', 'is_public', 'owner', 'hash', 'size', 'date', 'title', 'description', 'category', 'labels']
         resp.media = {key: dataset_meta_dict[key] for key in result_keys if key in dataset_meta_dict}
     
     def get_description(self, req, resp):
@@ -105,7 +101,12 @@ class DatasetResource:
             )
 
         resp.status = falcon.HTTP_200
-        resp.media = {'id': dataset_meta.id}
+        resp.media = {
+            'id': dataset_meta.id,
+            'date': dataset_meta.base.date,
+            'size': dataset_meta.base.size,
+            'hash': dataset_meta.base.hash
+        }
 
     def dataset_already_uploaded(self, req, resp, id):
         logger.debug('Dataset {id} alerady uploaded'.format(id=id))
@@ -115,22 +116,22 @@ class DatasetResource:
             description="Dataset already uploaded")
 
     def upload_dataset(self, req, resp, id):
+        user_id = req.context['user']
+        logger.debug('Authorize user {id}'.format(id=user_id))
+
         try:
-            dataset_meta = metadata.DatasetMetadata.from_id(id=id)
+            context = {'user_id': user_id}
+            dataset_meta = manager.get_dataset(id, context)
         except metadata.DoesNotExist:
-            dataset_meta = None
-
-        if dataset_meta:
-            if dataset_meta.status == metadata.dataset.PENDING:
-                self.save_dataset(req, resp, dataset_meta)
-
-            elif dataset_meta.status == metadata.dataset.RECEIVED:
-                self.dataset_already_uploaded(req, resp, id)
-        else:
             raise falcon.HTTPNotFound(
                 title="Dataset not found",
                 description="Dataset metadata does not exist"
             )
+
+        if dataset_meta.status == metadata.dataset.PENDING:
+            self.save_dataset(req, resp, dataset_meta)
+        elif dataset_meta.status == metadata.dataset.RECEIVED:
+            self.dataset_already_uploaded(req, resp, id)
 
     @jsonschema.validate(DATASET_SCHEMA)
     def create_dataset_meta(self, req, resp):